Separate parse and compile code
